+++ app/src/main/java/com/techcourse/handlermapping/CombinedHandlerMapping.java
@@ -1,7 +1,8 @@
-package com.techcourse;
+package com.techcourse.handlermapping;
 
 import com.interface21.webmvc.servlet.mvc.asis.Controller;
 import com.interface21.webmvc.servlet.mvc.tobe.AnnotationHandlerMapping;
+import com.techcourse.HandlerMapping;
 import jakarta.servlet.http.HttpServletRequest;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
