+++ app/src/main/java/com/techcourse/handlermapping/ManualHandlerMapping.java
@@ -1,4 +1,4 @@
-package com.techcourse;
+package com.techcourse.handlermapping;
 
 import com.interface21.webmvc.servlet.mvc.asis.Controller;
 import com.interface21.webmvc.servlet.mvc.asis.ForwardController;
